Fixed a bunch of shit, hopefully.

+++ distro_packages/arch.sh
@@ -1,22 +1,24 @@
 ### Run as root
 #! /bin/bash
 sudo sh -c "
-  pacman -Syu;
-  pacman -S
-    facter		\
-    libnotify-bin	\
-    nmap		\
-    ntp			\
-    python		\
-    python-dev		\
-    python-pip		\
-    python-venv		\
-    sl			\
-    snapd		\
-    speedtest-cli	\
-    thefuck		\
-    tldr		\
-    lm-sensors
+	pacman -Syu;
+	pacman -S
+		facter		\
+		libnotify-bin	\
+		lm-sensors	\
+		nmap		\
+		ntp		\
+		python		\
+		python-dev	\
+		python-pip	\
+		python-venv	\
+		sl		\
+		snapd		\
+		speedtest-cli	\
+		thefuck		\
+		tldr		\
+		vim
 "
-
-sudo ln -s /var/lib/snapd/snap /snap
+if [ ! -L /snap ]; then
+	sudo ln -s /var/lib/snapd/snap /snap
+fi
